Changed success message and email language, buttons components put into folder. Api routes organised to one folder
Closes #114
